Refactored common piece functionality into new abstract base class Piece.
